requirements.txt
# Metapod - Autonomous Backend Refactor & Hardening Agent
# Core dependencies - keeping minimal for maximum compatibility

# HTTP client for research
aiohttp>=3.9.0
requests>=2.31.0

# HTML parsing for research
beautifulsoup4>=4.12.0
lxml>=4.9.0

# YAML configuration
PyYAML>=6.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code analysis
ast-decompiler>=1.0.0

# Optional: Advanced dependencies for specific stacks
# Uncomment based on your target stack:

# Python stack
# fastapi>=0.104.0
# pydantic>=2.4.0
# sqlalchemy>=2.0.0
# structlog>=23.1.0

# Development tools
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.5.0
# flake8>=6.0.0
